# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Fiber Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5001",
    "basePath": "/",
    "paths": {
        "/admin/email": {
            "put": {
                "description": "Update email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "update email",
                "parameters": [
                    {
                        "description": "Email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "create a new email",
                "parameters": [
                    {
                        "description": "Email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Email"
                        }
                    }
                }
            }
        },
        "/admin/email/{company_id}": {
            "get": {
                "description": "Get email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Get email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Company ID",
                        "name": "company_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Email"
                        }
                    }
                }
            }
        },
        "/admin/email/{id}": {
            "delete": {
                "description": "Delete email by given ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "delete email by given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Email ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/emails": {
            "get": {
                "description": "Get emails",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Get emails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Email"
                        }
                    }
                }
            }
        },
        "/admin/notice": {
            "put": {
                "description": "Update notice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notice"
                ],
                "summary": "update notice",
                "parameters": [
                    {
                        "description": "Notice",
                        "name": "notice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new notice.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notice"
                ],
                "summary": "create a new notice",
                "parameters": [
                    {
                        "description": "Notice",
                        "name": "notice",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Notice"
                        }
                    }
                }
            }
        },
        "/admin/notice/{id}": {
            "get": {
                "description": "Get notice by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get a notice by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Notice"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete notice by given ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notice"
                ],
                "summary": "delete notice by given ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/admin/notices": {
            "get": {
                "description": "Get notices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get notices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Notice"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Email": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "email_domain": {
                    "type": "string"
                },
                "email_valid_id": {
                    "type": "integer"
                }
            }
        },
        "models.Notice": {
            "description": "공지사항 모델",
            "type": "object",
            "required": [
                "profile_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "notice_contents": {
                    "type": "string",
                    "example": "notice_contents"
                },
                "notice_id": {
                    "type": "integer"
                },
                "notice_title": {
                    "type": "string",
                    "example": "notice_title"
                },
                "profile_id": {
                    "type": "string",
                    "example": "profile_id"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}